requirements.txt
albumentations==1.3.0
einops==0.6.0
fastai==2.7.10
joblib==1.2.0
numpy
nvidia_dali_cuda110
opencv_python_headless==4.6.0.66
pandas
Pillow==9.4.0
pydicom==2.3.1
pytorch_lightning==1.8.3.post1
scikit_learn
timm==0.6.12
tqdm==4.64.1
